refactored the group code
