+++ pkg/groups/manager.go
@@ -1,42 +1,12 @@
-// Package groups provides functionality for managing logical groupings of MCP servers.
-// It includes types and interfaces for creating, retrieving, listing, and deleting groups.
 package groups
 
 import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"os"
 
 	"github.com/stacklok/toolhive/pkg/state"
 )
-
-// Group represents a logical grouping of MCP servers.
-type Group struct {
-	Name string `json:"name"`
-	// MCP servers will be added in a followup story.
-}
-
-// WriteJSON serializes the Group to JSON and writes it to the provided writer
-func (g *Group) WriteJSON(w *os.File) error {
-	encoder := json.NewEncoder(w)
-	encoder.SetIndent("", "  ")
-	return encoder.Encode(g)
-}
-
-// Manager provides operations for managing groups
-type Manager interface {
-	// Create creates a new group with the given name
-	Create(ctx context.Context, name string) error
-	// Get retrieves a group by name
-	Get(ctx context.Context, name string) (*Group, error)
-	// List returns all groups
-	List(ctx context.Context) ([]*Group, error)
-	// Delete removes a group by name
-	Delete(ctx context.Context, name string) error
-	// Exists checks if a group exists
-	Exists(ctx context.Context, name string) (bool, error)
-}
 
 // manager implements the Manager interface
 type manager struct {
